extract gcp data in settings, add description, and delete UserViewSet
