+++ .gitignore
@@ -76,5 +76,3 @@ media/
 
 # Venv
 venv
-
-backend/env_variables.yaml
