+++ backend/config/settings.py
@@ -9,8 +9,19 @@ https://docs.djangoproject.com/en/dev/ref/settings/
 """
 import environ
 import os
-from config.settings_gcp import get_setting
 from datetime import timedelta
+if os.getenv('GAE_APPLICATION', None):
+    from google.cloud import datastore
+
+    client = datastore.Client()
+    # The datastore extracted here was previously defined manually in 
+    # the Google Cloud Platform web console
+    def get_setting(name):
+        query = client.query(kind="Settings", filters=[["name", "=", name]])
+        entitys = list(query.fetch())
+        if not entitys:
+            raise ValueError("Could not find any entity")
+        return entitys[0]['value']
 
 ROOT_DIR = environ.Path(__file__) - 2
