+++ frontend/src/router.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import {vueAuth} from './main'
-
+import axios from 'axios'
 import HomeComponent from '@/components/HomeComponent.vue'
 import NavbarComponent from '@/components/NavbarComponent.vue'
 import FailComponent from '@/components/FailComponent.vue'
